requirements.txt
google-cloud-vision
google-cloud-translate==2.0.1
google-generativeai
python-dotenv
